refactor: 마켓 - 템플릿으로 패키지명 변경
