+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/home/HomeActivity.kt
@@ -9,8 +9,8 @@ import com.bottari.presentation.R
 import com.bottari.presentation.common.base.BaseActivity
 import com.bottari.presentation.databinding.ActivityHomeBinding
 import com.bottari.presentation.view.home.bottari.BottariFragment
-import com.bottari.presentation.view.home.market.TemplateFragment
 import com.bottari.presentation.view.home.profile.ProfileFragment
+import com.bottari.presentation.view.home.template.TemplateFragment
 import com.bottari.presentation.view.setting.SettingActivity
 
 class HomeActivity : BaseActivity<ActivityHomeBinding>(ActivityHomeBinding::inflate) {
