+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/home/template/adapter/TemplateAdapter.kt
@@ -1,10 +1,10 @@
-package com.bottari.presentation.view.home.market.adapter
+package com.bottari.presentation.view.home.template.adapter
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import com.bottari.presentation.model.BottariTemplateUiModel
-import com.bottari.presentation.view.home.market.listener.OnTemplateClickListener
+import com.bottari.presentation.view.home.template.listener.OnTemplateClickListener
 
 class TemplateAdapter(
     private val clickListener: OnTemplateClickListener,
