+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/home/template/listener/OnTemplateClickListener.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.view.home.market.listener
+package com.bottari.presentation.view.home.template.listener
 
 fun interface OnTemplateClickListener {
     fun onTemplateClick(bottariTemplateId: Long)
